break up celltypestable component into multiple files
